Add execution metadata to PATCH endpoint
