+++ backend/test_observer/controllers/test_executions/patch.py
@@ -25,6 +25,15 @@ from test_observer.data_access.setup import get_db
 
 from .models import TestExecutionsPatchRequest
 
+from sqlalchemy import tuple_
+from sqlalchemy.dialects.postgresql import insert as pg_insert
+
+from test_observer.data_access.models import (
+    TestExecutionMetadata,
+    test_execution_metadata_association_table,
+)
+from test_observer.controllers.execution_metadata.models import ExecutionMetadata
+
 router = APIRouter()
 
 
@@ -49,9 +58,13 @@ def patch_test_execution(
     if request.ci_link is not None:
         test_execution.ci_link = str(request.ci_link)
 
+    if request.execution_metadata is not None:
+        _add_execution_metadata(test_execution, request.execution_metadata, db)
+
     _set_test_execution_status(request, test_execution)
 
     db.commit()
+    db.refresh(test_execution)
     return test_execution
 
 
@@ -69,3 +82,60 @@ def _set_test_execution_status(
             test_execution.status = TestExecutionStatus.FAILED
         case ("COMPLETED", _):
             test_execution.status = TestExecutionStatus.PASSED
+
+
+def _add_execution_metadata(
+    test_execution: TestExecution,
+    execution_metadata: ExecutionMetadata,
+    db: Session,
+) -> None:
+    # Unpack metadata into list of tuples
+    execution_metadata_rows = execution_metadata.to_rows()
+
+    # Exit if none given
+    if len(execution_metadata_rows) == 0:
+        return
+
+    # Create any missing execution metadata
+    db.execute(
+        pg_insert(TestExecutionMetadata)
+        .values(
+            [
+                {
+                    "category": execution_metadata.category,
+                    "value": execution_metadata.value,
+                }
+                for execution_metadata in execution_metadata_rows
+            ]
+        )
+        .on_conflict_do_nothing()
+    )
+
+    # Fetch all execution metadata ids
+    execution_metadata_ids = (
+        db.query(TestExecutionMetadata.id)
+        .filter(
+            tuple_(TestExecutionMetadata.category, TestExecutionMetadata.value).in_(
+                [
+                    (execution_metadata.category, execution_metadata.value)
+                    for execution_metadata in execution_metadata_rows
+                ]
+            )
+        )
+        .all()
+    )
+
+    # Attach any missing execution metadata to the test execution
+    db.execute(
+        pg_insert(test_execution_metadata_association_table)
+        .values(
+            [
+                {
+                    "test_execution_id": test_execution.id,
+                    "test_execution_metadata_id": execution_metadata_id[0],
+                }
+                for execution_metadata_id in execution_metadata_ids
+            ]
+        )
+        .on_conflict_do_nothing()
+    )
